+++ src/api/endpoints/mod.rs
@@ -7,3 +7,4 @@ pub(in crate::api) mod web_index;
 pub(in crate::api) mod hls_api;
 mod user_api;
 pub(in crate::api) mod hdhomerun_api;
+mod api_playlist_utils;