+++ src/api/endpoints/v1_api.rs
@@ -7,21 +7,18 @@ use crate::auth::access_token::create_access_token;
 use crate::auth::authenticator::validator_admin;
 use crate::m3u_filter_error::M3uFilterError;
 use crate::model::api_proxy::{ApiProxyConfig, ApiProxyServerInfo, ProxyUserCredentials, TargetUser};
-use crate::model::config::{validate_targets, Config, ConfigDto, ConfigInput, ConfigInputOptions, ConfigSource, ConfigTarget, InputType, TargetType};
-use crate::model::playlist::{M3uPlaylistItem, PlaylistGroup, PlaylistItemType, XtreamCluster, XtreamPlaylistItem};
+use crate::model::config::{validate_targets, Config, ConfigDto, ConfigInput, ConfigInputOptions, ConfigSource, ConfigTarget, InputType};
+use crate::model::playlist::{XtreamPlaylistItem};
 use crate::processing::processor::playlist;
 use crate::repository::user_repository::store_api_user;
-use crate::repository::{m3u_repository, xtream_repository};
 use crate::utils::file::config_reader;
-use crate::utils::file::file_lock_manager::FileReadGuard;
 use crate::utils::network::request::sanitize_sensitive_info;
-use crate::utils::network::{m3u, xtream};
 use axum::response::IntoResponse;
 use log::error;
-use serde::Serialize;
-use serde_json::{json, Value};
-use std::collections::{HashMap, HashSet};
+use serde_json::{json};
+use std::collections::{HashSet};
 use std::sync::Arc;
+use crate::api::endpoints::api_playlist_utils::{get_playlist, get_playlist_for_target};
 
 fn intern_save_config_api_proxy(backup_dir: &str, api_proxy: &ApiProxyConfig, file_path: &str) -> Option<M3uFilterError> {
     match config_reader::save_api_proxy(file_path, backup_dir, api_proxy) {
@@ -180,184 +177,11 @@ fn create_config_input_for_xtream(username: &str, password: &str, host: &str) ->
     }
 }
 
-#[derive(serde::Serialize, serde::Deserialize)]
-struct PlaylistResponseGroup {
-    id: u32,
-    title: String,
-    channels: serde_json::Value,
-    xtream_cluster: XtreamCluster,
-}
-
-#[derive(serde::Serialize, serde::Deserialize)]
-struct PlaylistResponse {
-    live: Option<Vec<PlaylistResponseGroup>>,
-    vod: Option<Vec<PlaylistResponseGroup>>,
-    series: Option<Vec<PlaylistResponseGroup>>,
-}
-
-fn group_playlist_items<T>(
-    cluster: XtreamCluster,
-    iter: impl Iterator<Item=T>,
-    get_group: fn(&T) -> String,
-) -> Vec<PlaylistResponseGroup>
-where
-    T: Serialize,
-{
-    let mut groups: HashMap<String, Vec<T>> = HashMap::new();
-
-    for item in iter {
-        let group_key = get_group(&item);
-        groups.entry(group_key)
-            .or_default()
-            .push(item);
-    }
-
-    groups
-        .into_iter()
-        .enumerate()
-        .map(|(index, (key, value))| PlaylistResponseGroup {
-            #[allow(clippy::cast_possible_truncation)]
-            id: index as u32,
-            title: key.to_string(),
-            channels: serde_json::to_value(value).unwrap_or(Value::Null),
-            xtream_cluster: cluster,
-        })
-        .collect()
-}
-
-fn group_playlist_items_by_cluster(params: Option<(FileReadGuard,
-                                                   impl Iterator<Item=(M3uPlaylistItem, bool)>)>) ->
-                                   (Vec<M3uPlaylistItem>, Vec<M3uPlaylistItem>, Vec<M3uPlaylistItem>) {
-    if params.is_none() {
-        return (vec![], vec![], vec![]);
-    }
-    let mut live = Vec::new();
-    let mut video = Vec::new();
-    let mut series = Vec::new();
-    let (guard, iter) = params.unwrap();
-    for (item, _) in iter {
-        match item.item_type {
-            PlaylistItemType::Live
-            | PlaylistItemType::Catchup
-            | PlaylistItemType::LiveUnknown
-            | PlaylistItemType::LiveHls
-            | PlaylistItemType::LiveDash => {
-                live.push(item);
-            }
-            PlaylistItemType::Video => {
-                video.push(item);
-            }
-            PlaylistItemType::Series
-            | PlaylistItemType::SeriesInfo => {
-                series.push(item);
-            }
-        }
-    }
-
-    drop(guard);
-
-    (live, video, series)
-}
-
-fn group_playlist_groups_by_cluster(playlist: Vec<PlaylistGroup>, input_type: InputType) -> (Vec<PlaylistResponseGroup>, Vec<PlaylistResponseGroup>, Vec<PlaylistResponseGroup>) {
-    let mut live = Vec::new();
-    let mut video = Vec::new();
-    let mut series = Vec::new();
-    for group in playlist {
-        let channels = group.channels.iter().map(|item| if input_type == InputType::M3u { serde_json::to_value(item.to_m3u()).unwrap() } else { serde_json::to_value(item.to_xtream()).unwrap() }).collect();
-        let grp = PlaylistResponseGroup {
-            id: group.id,
-            title: group.title,
-            channels,
-            xtream_cluster: group.xtream_cluster,
-        };
-        match group.xtream_cluster {
-            XtreamCluster::Live => live.push(grp),
-            XtreamCluster::Video => video.push(grp),
-            XtreamCluster::Series => series.push(grp),
-        }
-    }
-    (live, video, series)
-}
-
-async fn get_playlist(client: Arc<reqwest::Client>, cfg_input: Option<&ConfigInput>, cfg: &Config) -> impl IntoResponse + Send {
-    match cfg_input {
-        Some(input) => {
-            let (result, errors) =
-                match input.input_type {
-                    InputType::M3u => m3u::get_m3u_playlist(client, cfg, input, &cfg.working_dir).await,
-                    InputType::Xtream => xtream::get_xtream_playlist(client, input, &cfg.working_dir).await,
-                    InputType::M3uBatch | InputType::XtreamBatch => (vec![], vec![])
-                };
-            if result.is_empty() {
-                let error_strings: Vec<String> = errors.iter().map(std::string::ToString::to_string).collect();
-                (axum::http::StatusCode::BAD_REQUEST, axum::Json(json!({"error": error_strings.join(", ")}))).into_response()
-            } else {
-                let (live, vod, series) = group_playlist_groups_by_cluster(result, input.input_type);
-                let response = PlaylistResponse {
-                    live: Some(live),
-                    vod: Some(vod),
-                    series: Some(series),
-                };
-                (axum::http::StatusCode::OK, axum::Json(response)).into_response()
-            }
-        }
-        None => (axum::http::StatusCode::BAD_REQUEST, axum::Json(json!({"error": "Invalid Arguments"}))).into_response(),
-    }
-}
-
-// async fn get_categories_content(action: Result<(Option<PathBuf>, Option<String>), std::io::Error>) -> Option<String> {
-//     if let Ok((Some(file_path), _content)) = action {
-//         if let Ok(content) = tokio::fs::read_to_string(&file_path).await {
-//             // TODO deserialize like sax parser
-//             if let Ok(categories) = serde_json::from_str::<Vec<PlaylistXtreamCategory>>(&content) {
-//                 return serde_json::to_string(&categories).ok();
-//             }
-//         }
-//     }
-//     None
-// }
-
-async fn get_playlist_for_target(cfg_target: Option<&ConfigTarget>, cfg: &Arc<Config>) -> impl axum::response::IntoResponse + Send {
-    if let Some(target) = cfg_target {
-        if target.has_output(&TargetType::Xtream) {
-            // let live_categories = get_categories_content(xtream_repository::xtream_get_collection_path(cfg, target_name, xtream_repository::COL_CAT_LIVE)).await;
-            // let vod_categories = get_categories_content(xtream_repository::xtream_get_collection_path(cfg, target_name, xtream_repository::COL_CAT_VOD)).await;
-            // let series_categories = get_categories_content(xtream_repository::xtream_get_collection_path(cfg, target_name, xtream_repository::COL_CAT_SERIES)).await;
-
-            let live_channels = xtream_repository::iter_raw_xtream_playlist(cfg, target, XtreamCluster::Live).await
-                .map(|(_guard, iter)| group_playlist_items(XtreamCluster::Live, iter.map(|(v, _)| v), |item| item.group.clone()));
-            let vod_channels = xtream_repository::iter_raw_xtream_playlist(cfg, target, XtreamCluster::Video).await
-                .map(|(_guard, iter)| group_playlist_items(XtreamCluster::Video, iter.map(|(v, _)| v), |item| item.group.clone()));
-            let series_channels = xtream_repository::iter_raw_xtream_playlist(cfg, target, XtreamCluster::Series).await
-                .map(|(_guard, iter)| group_playlist_items(XtreamCluster::Series, iter.map(|(v, _)| v), |item| item.group.clone()));
-
-            let response = PlaylistResponse {
-                live: live_channels,
-                vod: vod_channels,
-                series: series_channels,
-            };
-
-            return (axum::http::StatusCode::OK, axum::Json(response)).into_response();
-        } else if target.has_output(&TargetType::M3u) {
-            let all_channels = m3u_repository::iter_raw_m3u_playlist(cfg, target).await;
-            let (live_channels, vod_channels, series_channels) = group_playlist_items_by_cluster(all_channels);
-            let response = PlaylistResponse {
-                live: Some(group_playlist_items(XtreamCluster::Live, live_channels.into_iter(), |item| item.group.clone())),
-                vod: Some(group_playlist_items(XtreamCluster::Video, vod_channels.into_iter(), |item| item.group.clone())),
-                series: Some(group_playlist_items(XtreamCluster::Series, series_channels.into_iter(), |item| item.group.clone())),
-            };
-
-            return (axum::http::StatusCode::OK, axum::Json(response)).into_response();
-        }
-    }
-    (axum::http::StatusCode::BAD_REQUEST, axum::Json(json!({"error": "Invalid Arguments"}))).into_response()
-}
 
 async fn playlist_content(
     axum::extract::State(app_state): axum::extract::State<Arc<AppState>>,
     axum::extract::Json(playlist_req): axum::extract::Json<PlaylistRequest>,
-) -> impl axum::response::IntoResponse + Send {
+) -> impl IntoResponse + Send {
     match playlist_req.rtype {
         PlaylistRequestType::Input => {
             if let Some(source_id) = playlist_req.source_id {